refactor(auth): rename driveAuth.js to driveClient.js for clarity
